Start of a bare bones wiki that is a client of the socialtext REST API. A Turbogears APP based the Turbogears 20 minute wiki tutorial.
Requires changes to the python socialtext.rest.simple_client
code, which will be checked in tomorrow.

This is working, but has some configuration details
that need polish, needs style (don't they always),
and generally needs a second day.

This was cdent and matthew's wikithon thing. Not much actual
coding was done today because there's sickness.

Most of this stuff is auto-generated code. See st/model.py and
st/controllers.py for the meat.

Was fun.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext-clients/trunk@130 4ee4aff5-2028-0410-b109-97ae9d850f25