This was all abstracted from our Buildbot's master.cfg file.
We override quite a lot of functionality, and getting it out of a single
module helps quite a lot. With the exception of socialtext.buildbot.common,
the various overridden subclasses follow the same code hierarchy as in
Buildbot's source tree. Hopefully, this will make it easier to explore and
extend.

git-svn-id: https://repo.socialtext.net:8999/svn/socialtext-clients/trunk@245 4ee4aff5-2028-0410-b109-97ae9d850f25