Add a decorator that lets you declare an action method as needed access control.
The decorator effectively wraps the method in another,
only calling it if the incoming request has the right
stuff.

One of the reasons for doing this as a decorator is
that they are stackable, and I find them _incredibly_
readable (if you accept that the decorating code is okay).

The idea here is than only an auth'd client (with the 
special closet cookie) should be allowed to write to
the store. Since to GET anything from the store requires
a hard to guess UUID there's no need to do read 
access control (yet).


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext-clients/trunk@299 4ee4aff5-2028-0410-b109-97ae9d850f25