Add cache-handling to closet so the client can do, if it wants, caching. Since the closet is a worm, this is really easy to do: the UUID URI resource can act as the the Etag validator for the resource and we can set the max-age on the cache for as long as we choose in closet.cache_control (see close.py).
The "test" doesn't really work as a test: it proves itself
by the human observation of the logs that happen when the
test runs. Basically you see that we don't do a get because
we have it cached on the client side, then we expire the cache
and do send a real get, with an Etag, and get a 304 back,
so no file read is required.

HTTP is freaking awesome.

Note that this is HTTP/1.1 cache handling.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext-clients/trunk@305 4ee4aff5-2028-0410-b109-97ae9d850f25