Move config information into a YAML file.
This just gets it started, is not meant to be super 
tidy or anything. The config as done right now will
not allow the poster to be on a separate server from
the getter and putter. That will be fixed.

Theoretically the getter, putter and poster could all
be on separate machines, but the getter and putter
need to have access to the same (potentially virtual)
disk space.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext-clients/trunk@306 4ee4aff5-2028-0410-b109-97ae9d850f25